Dynamic filtering using static sites
Webflow is an incredibly flexible tool. And while it doesn’t offer a native solution to every problem, custom code makes it possible to implement almost anything your audience might need — including the ability to find what they need. Whether it’s a blog, a portfolio, or an ecommerce project, people want to find what they need quickly and easily. With that in mind, I’m going to show you how to integrate the MixitUp JavaScript library so your audience can filter and sort content.

Why MixItUp3?

MixItUp3 is a dependency-free, client-side JavaScript library for animated filtering and sorting. It’s great for managing content that needs to be categorized or ordered like portfolios, galleries, stores, and blogs. It can be used with both static and dynamic content. 

Anything you can display on the page can be categorized, filtered, and sorted by: 

DatePriceCategoryTagA combination of the above

When a website visitor selects a filter, MixitUp shows elements with the class name specified in the filter settings and hides the rest.

The library is powerful and includes a variety of features. The basic functionality of MixitUp is open-source and free to use for personal projects. If you’re using it for commercial projects, you’ll need to buy a license with a one-time, reasonably priced fee.

In this tutorial we’ll look at the most common uses for MixItUp3:

Filtering content on a simple, static page Filtering and sorting dynamic posts on a CMS-powered blog page
1. Filtering content on a simple, static page
The free Webflow template Incredible is a great fit for our first example. Let's filter the template’s photo gallery.

Step 1: Creating a layout

By default, MixItUp will check the HTML structure of the page, look for a div with the class name we’ll specify in the Designer and in the code, and filter all elements that match the class name mix inside that div. 

So, our layout should contain a div with items with the class mix. Let’s call that div mix-container 

Note: The template we’re using for this tutorial already has some existing class names — sometimes we’ll use combo classes or rename existing classes which will work just fine for filtering purposes. In your original project, you can create all class names and styles.

Step 2: Applying filtering categories

To let MixItUp filter elements, each element should have a class name that specifies the category it belongs to. We’ll apply those selectors as a combo-classes to each mix element. 

For our photo gallery, these combo classes will be types of photos:

black-and-whitemacronaturepeopletravelStep 3: Creating filter controls

You can make filter controls out of different types of HTML elements like buttons, links, or input elements (checkboxes, selects etc.). In this case, buttons work nicely. 

Let’s create a filter-controls div and put all our filter-button elements inside it. We need one button for each category, plus one to let the user clear all filters.